feat(line-chart-web): remove duplicate file and move playground utils to appropriate folder
